+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.13
+
+* **bugfix**: A RUNTIME parent does not display DEBUG child's metric